requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
google-cloud-dialogflow==2.19.1
google-cloud-vertexai==1.7.0
protobuf==3.20.0
